% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Data Generating Process for Spatial Scalar-on-Function Model}
\usage{
data_generation(n, j, rho, mean.e, sig.e, out.p = 0)
}
\arguments{
\item{n}{Integer. Sample size.}

\item{j}{Integer. Number of grid points.}

\item{rho}{Numeric. Spatial autoregressive parameter.}

\item{mean.e}{Numeric. Mean of error terms.}

\item{sig.e}{Numeric. Standard deviation of error Terms.}

\item{out.p}{Numeric. Proportion of outliers to introduce (default is 0).}
}
\value{
A list with the following elements:
\describe{
\item{y}{Response vector of length \code{n}.}
\item{x}{Matrix of functional predictor values.}
\item{w}{Spatial weight matrix of size \code{n x n}.}
\item{tcoefs}{True coefficient function used to generate the data.}
\item{out.index}{Indices of outlier observations, if any.}
}
}
\description{
Generates synthetic data from a spatial scalar-on-function regression model.
}
\examples{
\dontrun{
sim_data <- data_generation(n=250, j=101, rho=0.5, mean.e=1, sig.e=1, out.p = 0)
y <- sim_data$y
x <- sim_data$x
w <- sim_data$w
tcoef <- sim_data$tcoefs
out.index <- sim_data$out.index
}

}
