% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rssofqr.R
\name{predict_rssofqr}
\alias{predict_rssofqr}
\title{Predict Method for Robust Functional Quantile Regression with Spatial Effects}
\usage{
predict_rssofqr(object, xnew, wnew)
}
\arguments{
\item{object}{A fitted model object returned by \code{\link{rssofqr}}.}

\item{xnew}{Numeric matrix of size \code{n x p}. New functional predictor data (on the same grid used in training).}

\item{wnew}{Numeric matrix of size \code{n x n}. Spatial weight matrix for the new observations.}
}
\value{
A numeric vector of predicted values of length \code{n}.
}
\description{
Makes predictions for new observations using a fitted robust spatial functional quantile regression model.
}
\examples{
\dontrun{
sim_data <- data_generation(n=250, j=101, rho=0.5, mean.e=1, sig.e=1, out.p = 0)
y <- sim_data$y
x <- sim_data$x
w <- sim_data$w
fit_kim <- rssofqr(y=y, x=x, w=w, tau=tau, model = "KM", CV="TRUE")
fit_ch <- rssofqr(y=y, x=x, w=w, tau=tau, model = "CH", CV="TRUE")
sim_test <- data_generation(n=250, j=101, rho=0.5, mean.e=1, sig.e=1, out.p = 0)
y_test <- sim_test$y
x_test <- sim_test$x
w_test <- sim_test$w
predict_kim <- predict_rssofqr(object = fit_kim, xnew = x_test, wnew = w_test)
predict_ch <- predict_rssofqr(object = fit_ch, xnew = x_test, wnew = w_test)
}

}
