% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssofqr.R
\name{rssofqr}
\alias{rssofqr}
\title{Robust Functional Quantile Regression Model with Spatial Autocorrelation}
\usage{
rssofqr(
  y,
  x,
  w,
  tau,
  tau_comp = NULL,
  gp = NULL,
  nbasis = NULL,
  h = NULL,
  model = c("KM", "CH"),
  CV = c("TRUE", "FALSE")
)
}
\arguments{
\item{y}{Numeric vector of length \code{n}. Response variable.}

\item{x}{Numeric matrix of size \code{n x p}. Functional predictor observed on a common grid.}

\item{w}{Numeric matrix of size \code{n x n}. Spatial weight matrix.}

\item{tau}{Numeric. Quantile level to estimate (between 0 and 1).}

\item{tau_comp}{Numeric. Quantile level to estimate for extracted components (between 0 and 1).}

\item{gp}{Numeric vector of length \code{p}. Grid of argument values for the functional predictor. If \code{NULL}, defaults to a regular grid between 0 and 1.}

\item{nbasis}{Integer. Number of basis functions to use in RFPQR. If \code{NULL}, defaults to \code{min(10, p / 4)}.}

\item{h}{Integer. Number of extracted components. If \code{NULL}, it is optimally determined when CV is True.}

\item{model}{Character string. Estimation method to use. One of \code{"KM"} (Kim et al.) or \code{"CH"} (Chernozhukov et al.).}

\item{CV}{Character string. Determine the number of components to extract based on the BIC. One of \code{"TRUE"} or \code{"FALSE"}.}
}
\value{
A list containing:
\describe{
\item{b}{Estimated coefficient vector in RFPQR space.}
\item{b0}{Estimated intercept.}
\item{bhat}{Estimated coefficient function in the original function space.}
\item{rho}{Estimated spatial autoregressive parameter.}
\item{opt.h}{Number of extracted optimal components.}
\item{opt.nbasis}{Number of basis functions.}
\item{fitted.values}{Fitted values from the model.}
\item{residuals}{Residuals from the model.}
\item{tau}{Quantile level used.}
\item{details}{Output from \code{getRFPQR}, including RFPQR scores and basis functions.}
}
}
\description{
Fits a spatial scalar-on-function quantile regression model using either the Kim and Muller (2004)
or Chernozhukov and Hansen (2006) method.
}
\examples{
\dontrun{
sim_data <- data_generation(n=250, j=101, rho=0.5, mean.e=1, sig.e=1, out.p = 0)
y <- sim_data$y
x <- sim_data$x
w <- sim_data$w
fit_kim <- rssofqr(y=y, x=x, w=w, tau=tau, model = "KM", CV="TRUE")
fit_ch <- rssofqr(y=y, x=x, w=w, tau=tau, model = "CH", CV="TRUE")
}

}
